feat: centralize styles
