+++ src/components/HomeContent.tsx
@@ -1,65 +1,66 @@
-import React from 'react';
-import {
-  SafeAreaView,
-  ScrollView,
-  StatusBar,
-  StyleSheet,
-  Text,
-  useColorScheme,
-  View,
-} from 'react-native';
-import {
-  Colors,
-  DebugInstructions,
-  Header,
-  LearnMoreLinks,
-  ReloadInstructions,
-} from 'react-native/Libraries/NewAppScreen';
-import Section from '../elements/Section';
+// import React from 'react';
+// import {
+//   SafeAreaView,
+//   ScrollView,
+//   StatusBar,
+//   StyleSheet,
+//   Text,
+//   useColorScheme,
+//   View,
+// } from 'react-native';
+// import {
+//   Colors,
+//   DebugInstructions,
+//   Header,
+//   LearnMoreLinks,
+//   ReloadInstructions,
+// } from 'react-native/Libraries/NewAppScreen';
+// import Section from '../elements/Section';
 
 
-const HomeContent = () => {
-  const isDarkMode = useColorScheme() === 'dark';
+// const HomeContent = () => {
+//   const isDarkMode = useColorScheme() === 'dark';
 
-  const backgroundStyle = {
-    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
-  };
+//   const backgroundStyle = {
+//     backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
+//   };
 
-  return (
-    <ScrollView
-      contentInsetAdjustmentBehavior="automatic"
-      style={backgroundStyle}>
-      {/* <Header /> */}
-      <View
-        style={{
-          backgroundColor: isDarkMode ? Colors.black : Colors.white,
-        }}>
-        <Section title="Step One">
-          <Text style={styles.highlight}>New Features</Text> in our new application.
-        </Section>
-        {/* <LearnMoreLinks /> */}
-      </View>
-    </ScrollView>
-  )
-}
+//   return (
+//     <ScrollView
+//       contentInsetAdjustmentBehavior="automatic"
+//       style={backgroundStyle}>
+//       {/* <Header /> */}
+//       <View
+//         style={{
+//           backgroundColor: isDarkMode ? Colors.black : Colors.white,
+//         }}>
+//         <Section title="Step One">
+//           <Text style={styles.highlight}>New Features</Text> in our new application.
+//         </Section>
+//         {/* <LearnMoreLinks /> */}
+//       </View>
+//     </ScrollView>
+//   )
+// }
 
-const styles = StyleSheet.create({
-  sectionContainer: {
-    marginTop: 32,
-    paddingHorizontal: 24,
-  },
-  sectionTitle: {
-    fontSize: 24,
-    fontWeight: '600',
-  },
-  sectionDescription: {
-    marginTop: 8,
-    fontSize: 18,
-    fontWeight: '400',
-  },
-  highlight: {
-    fontWeight: '700',
-  },
-});
+// const styles = StyleSheet.create({
+//   sectionContainer: {
+//     marginTop: 32,
+//     paddingHorizontal: 24,
+//   },
+//   sectionTitle: {
+//     fontSize: 24,
+//     fontWeight: '600',
+//   },
+//   sectionDescription: {
+//     marginTop: 8,
+//     fontSize: 18,
+//     fontWeight: '400',
+//   },
+//   highlight: {
+//     fontWeight: '700',
+//   },
+// });
 
-export default HomeContent;
+// export default HomeContent;
+export {};
