+++ src/elements/Section.tsx
@@ -11,6 +11,8 @@ import {
     View,
 } from 'react-native';
 
+import styles from '../styles';
+
 import {
     Colors,
     DebugInstructions,
@@ -48,24 +50,5 @@ function Section({ children, title }: SectionProps): JSX.Element {
         </View>
     );
 }
-
-const styles = StyleSheet.create({
-    sectionContainer: {
-        marginTop: 32,
-        paddingHorizontal: 24,
-    },
-    sectionTitle: {
-        fontSize: 24,
-        fontWeight: '600',
-    },
-    sectionDescription: {
-        marginTop: 8,
-        fontSize: 18,
-        fontWeight: '400',
-    },
-    highlight: {
-        fontWeight: '700',
-    },
-});
 
 export default Section;
